+++ test/deployables/Azorius-LinearERC20VotingWithHatsProposalCreation.test.ts
@@ -15,20 +15,20 @@ import {
   VotesERC20V1__factory,
   ModuleProxyFactory,
   GnosisSafeL2__factory,
-} from '../typechain-types';
+} from '../../typechain-types';
 
 import {
   getGnosisSafeL2Singleton,
   getGnosisSafeProxyFactory,
   getModuleProxyFactory,
-} from './GlobalSafeDeployments.test';
+} from '../global/GlobalSafeDeployments.test';
 import {
   calculateProxyAddress,
   predictGnosisSafeAddress,
   buildSafeTransaction,
   safeSignTypedData,
   buildSignatureBytes,
-} from './helpers';
+} from '../helpers';
 
 describe('LinearERC20VotingWithHatsProposalCreation', () => {
   // Deployed contracts
