+++ test/deployables/MultisigFreezeGuard-ERC20FreezeVoting.test.ts
@@ -2,46 +2,38 @@ import { SignerWithAddress } from '@nomicfoundation/hardhat-ethers/signers';
 import { expect } from 'chai';
 import hre, { ethers } from 'hardhat';
 import {
-  ERC721FreezeVotingV1,
-  ERC721FreezeVotingV1__factory,
+  VotesERC20V1,
+  VotesERC20V1__factory,
+  ERC20FreezeVotingV1,
+  ERC20FreezeVotingV1__factory,
   MultisigFreezeGuardV1,
   MultisigFreezeGuardV1__factory,
-  ERC20FreezeVotingV1__factory,
-  MockERC721,
-  MockERC721__factory,
-  AzoriusV1,
-  AzoriusV1__factory,
-  LinearERC721VotingV1,
-  LinearERC721VotingV1__factory,
   GnosisSafeL2__factory,
   GnosisSafeL2,
-} from '../typechain-types';
+} from '../../typechain-types';
 import {
   getGnosisSafeL2Singleton,
   getGnosisSafeProxyFactory,
   getModuleProxyFactory,
-} from './GlobalSafeDeployments.test';
+} from '../global/GlobalSafeDeployments.test';
 import {
   buildSignatureBytes,
   buildSafeTransaction,
   safeSignTypedData,
   predictGnosisSafeAddress,
   calculateProxyAddress,
-} from './helpers';
-import time from './time';
+} from '../helpers';
+import time from '../time';
 
 describe('Child Multisig DAO with Azorius Parent', () => {
   // Deployed contracts
   let gnosisSafe: GnosisSafeL2;
   let freezeGuardMastercopy: MultisigFreezeGuardV1;
   let freezeGuard: MultisigFreezeGuardV1;
-  let freezeVotingMastercopy: ERC721FreezeVotingV1;
-  let freezeVoting: ERC721FreezeVotingV1;
-  let mockNFT: MockERC721;
-  let linearERC721Voting: LinearERC721VotingV1;
-  let linearERC721VotingMastercopy: LinearERC721VotingV1;
-  let azoriusMastercopy: AzoriusV1;
-  let azorius: AzoriusV1;
+  let freezeVotingMastercopy: ERC20FreezeVotingV1;
+  let freezeVoting: ERC20FreezeVotingV1;
+  let votesERC20Mastercopy: VotesERC20V1;
+  let votesERC20: VotesERC20V1;
 
   // Wallets
   let deployer: SignerWithAddress;
@@ -50,20 +42,13 @@ describe('Child Multisig DAO with Azorius Parent', () => {
   let owner3: SignerWithAddress;
   let tokenVetoer1: SignerWithAddress;
   let tokenVetoer2: SignerWithAddress;
-  let vetoer1Ids: number[];
-  let vetoer2Ids: number[];
   let freezeGuardOwner: SignerWithAddress;
-  let mintNFTData: string;
 
   // Gnosis
   let createGnosisSetupCalldata: string;
 
   const threshold = 2;
   const saltNum = BigInt('0x856d90216588f9ffc124d1480a440e1c012c7a816952bc968d737bae5d4e139c');
-
-  async function mintNFT(contract: MockERC721, receiver: SignerWithAddress): Promise<void> {
-    await contract.connect(receiver).mint(receiver.address);
-  }
 
   beforeEach(async () => {
     [deployer, owner1, owner2, owner3, tokenVetoer1, tokenVetoer2, freezeGuardOwner] =
@@ -104,104 +89,45 @@ describe('Child Multisig DAO with Azorius Parent', () => {
     // eslint-disable-next-line camelcase
     gnosisSafe = GnosisSafeL2__factory.connect(predictedGnosisSafeAddress, deployer);
 
+    // Deploy token mastercopy
+    votesERC20Mastercopy = await new VotesERC20V1__factory(deployer).deploy();
+
     const abiCoder = new ethers.AbiCoder(); // encode data
-
-    // Deploy Mock NFT
-    mockNFT = await new MockERC721__factory(deployer).deploy();
-
-    // dish out NFTs
-    await mintNFT(mockNFT, tokenVetoer1);
-    await mintNFT(mockNFT, tokenVetoer2);
-    await mintNFT(mockNFT, tokenVetoer2);
-    vetoer1Ids = [0];
-    vetoer2Ids = [1, 2];
-
-    mintNFTData = mockNFT.interface.encodeFunctionData('mint', [deployer.address]);
-
-    // Deploy Azorius module
-    azoriusMastercopy = await new AzoriusV1__factory(deployer).deploy();
-
-    const azoriusSetupCalldata =
+    const votesERC20SetupData =
       // eslint-disable-next-line camelcase
-      AzoriusV1__factory.createInterface().encodeFunctionData('setUp', [
+      VotesERC20V1__factory.createInterface().encodeFunctionData('setUp', [
         abiCoder.encode(
-          ['address', 'address', 'address', 'address[]', 'uint32', 'uint32'],
+          ['string', 'string', 'address[]', 'uint256[]'],
           [
-            owner1.address,
-            await gnosisSafe.getAddress(),
-            await gnosisSafe.getAddress(),
-            [],
-            60, // timelock period in blocks
-            60, // execution period in blocks
+            'DCNT',
+            'DCNT',
+            [tokenVetoer1.address, tokenVetoer2.address, await gnosisSafe.getAddress()],
+            [500, 600, 1000],
           ],
         ),
       ]);
 
     await moduleProxyFactory.deployModule(
-      await azoriusMastercopy.getAddress(),
-      azoriusSetupCalldata,
+      await votesERC20Mastercopy.getAddress(),
+      votesERC20SetupData,
       '10031021',
     );
 
-    const predictedAzoriusAddress = await calculateProxyAddress(
+    const predictedVotesERC20Address = await calculateProxyAddress(
       moduleProxyFactory,
-      await azoriusMastercopy.getAddress(),
-      azoriusSetupCalldata,
+      await votesERC20Mastercopy.getAddress(),
+      votesERC20SetupData,
       '10031021',
     );
 
-    azorius = AzoriusV1__factory.connect(predictedAzoriusAddress, deployer);
+    votesERC20 = VotesERC20V1__factory.connect(predictedVotesERC20Address, deployer);
 
-    // Deploy Linear ERC721 Voting Mastercopy
-    linearERC721VotingMastercopy = await new LinearERC721VotingV1__factory(deployer).deploy();
+    // Vetoers delegate their votes to themselves
+    await votesERC20.connect(tokenVetoer1).delegate(tokenVetoer1.address);
+    await votesERC20.connect(tokenVetoer2).delegate(tokenVetoer2.address);
 
-    const linearERC721VotingSetupCalldata =
-      // eslint-disable-next-line camelcase
-      LinearERC721VotingV1__factory.createInterface().encodeFunctionData('setUp', [
-        abiCoder.encode(
-          [
-            'address',
-            'address[]',
-            'uint256[]',
-            'address',
-            'uint32',
-            'uint256',
-            'uint256',
-            'uint256',
-          ],
-          [
-            owner1.address, // owner
-            [await mockNFT.getAddress()], // NFT addresses
-            [500], // NFT weights
-            await azorius.getAddress(), // Azorius module
-            60, // voting period in blocks
-            2, // quorom threshold
-            2, // proposer threshold
-            500000, // basis numerator, denominator is 1,000,000, so basis percentage is 50% (simple majority)
-          ],
-        ),
-      ]);
-
-    await moduleProxyFactory.deployModule(
-      await linearERC721VotingMastercopy.getAddress(),
-      linearERC721VotingSetupCalldata,
-      '10031021',
-    );
-
-    const predictedlinearERC721VotingAddress = await calculateProxyAddress(
-      moduleProxyFactory,
-      await linearERC721VotingMastercopy.getAddress(),
-      linearERC721VotingSetupCalldata,
-      '10031021',
-    );
-
-    linearERC721Voting = LinearERC721VotingV1__factory.connect(
-      predictedlinearERC721VotingAddress,
-      deployer,
-    );
-
-    // Deploy ERC721FreezeVoting mastercopy contract
-    freezeVotingMastercopy = await new ERC721FreezeVotingV1__factory(deployer).deploy();
+    // Deploy ERC20FreezeVoting mastercopy contract
+    freezeVotingMastercopy = await new ERC20FreezeVotingV1__factory(deployer).deploy();
 
     // Initialize FreezeVoting contract
     const freezeVotingSetupData =
@@ -211,10 +137,10 @@ describe('Child Multisig DAO with Azorius Parent', () => {
           ['address', 'uint256', 'uint32', 'uint32', 'address'],
           [
             freezeGuardOwner.address,
-            501, // freeze votes threshold
+            1090, // freeze votes threshold
             10, // freeze proposal period
             200, // freeze period
-            await linearERC721Voting.getAddress(), // strategy address
+            await votesERC20.getAddress(),
           ],
         ),
       ]);
@@ -232,7 +158,7 @@ describe('Child Multisig DAO with Azorius Parent', () => {
       '10031021',
     );
 
-    freezeVoting = ERC721FreezeVotingV1__factory.connect(predictedFreezeVotingAddress, deployer);
+    freezeVoting = ERC20FreezeVotingV1__factory.connect(predictedFreezeVotingAddress, deployer);
 
     // Deploy FreezeGuard mastercopy contract
     freezeGuardMastercopy = await new MultisigFreezeGuardV1__factory(deployer).deploy();
@@ -300,12 +226,15 @@ describe('Child Multisig DAO with Azorius Parent', () => {
         signatureBytes,
       ),
     ).to.emit(gnosisSafe, 'ExecutionSuccess');
+
+    // Gnosis Safe received the 1,000 tokens
+    expect(await votesERC20.balanceOf(await gnosisSafe.getAddress())).to.eq(1000);
   });
 
   describe('FreezeGuard Functionality', () => {
     it('Freeze parameters correctly setup', async () => {
       // Frozen Params init correctly
-      expect(await freezeVoting.freezeVotesThreshold()).to.eq(501);
+      expect(await freezeVoting.freezeVotesThreshold()).to.eq(1090);
       expect(await freezeVoting.freezeProposalPeriod()).to.eq(10);
       expect(await freezeVoting.freezePeriod()).to.eq(200);
       expect(await freezeVoting.owner()).to.eq(freezeGuardOwner.address);
@@ -316,26 +245,26 @@ describe('Child Multisig DAO with Azorius Parent', () => {
       expect(await freezeVoting.freezeProposalCreatedBlock()).to.eq(0);
 
       // Vetoer 1 casts 500 freeze votes
-      await freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
+      await freezeVoting.connect(tokenVetoer1).castFreezeVote();
       expect(await freezeVoting.isFrozen()).to.eq(false);
       expect(await freezeVoting.freezeProposalVoteCount()).to.eq(500);
       const latestBlock = await hre.ethers.provider.getBlock('latest');
       expect(await freezeVoting.freezeProposalCreatedBlock()).to.eq(latestBlock!.number);
 
-      await freezeVoting
-        .connect(tokenVetoer2)
-        [
-          'castFreezeVote(address[],uint256[])'
-        ]([await mockNFT.getAddress(), await mockNFT.getAddress()], vetoer2Ids);
+      await freezeVoting.connect(tokenVetoer2).castFreezeVote();
       expect(await freezeVoting.isFrozen()).to.eq(true);
     });
 
     it('A transaction can be timelocked and executed', async () => {
+      // Create transaction to set the guard address
+      const tokenTransferData = votesERC20.interface.encodeFunctionData('transfer', [
+        deployer.address,
+        1000,
+      ]);
+
       const tx = buildSafeTransaction({
-        to: await mockNFT.getAddress(),
-        data: mintNFTData,
+        to: await votesERC20.getAddress(),
+        data: tokenTransferData,
         safeTxGas: 1000000,
         nonce: await gnosisSafe.nonce(),
       });
@@ -384,13 +313,20 @@ describe('Child Multisig DAO with Azorius Parent', () => {
         signatureBytes,
       );
 
-      expect(await mockNFT.balanceOf(deployer.address)).to.eq(1);
+      expect(await votesERC20.balanceOf(await gnosisSafe.getAddress())).to.eq(0);
+      expect(await votesERC20.balanceOf(deployer.address)).to.eq(1000);
     });
 
     it("A transaction cannot be executed if it hasn't yet been timelocked", async () => {
+      // Create transaction to set the guard address
+      const tokenTransferData = votesERC20.interface.encodeFunctionData('transfer', [
+        deployer.address,
+        1000,
+      ]);
+
       const tx = buildSafeTransaction({
-        to: await mockNFT.getAddress(),
-        data: mintNFTData,
+        to: await votesERC20.getAddress(),
+        data: tokenTransferData,
         safeTxGas: 1000000,
         nonce: await gnosisSafe.nonce(),
       });
@@ -418,9 +354,15 @@ describe('Child Multisig DAO with Azorius Parent', () => {
     });
 
     it("A transaction cannot be timelocked if the signatures aren't valid", async () => {
+      // Create transaction to set the guard address
+      const tokenTransferData = votesERC20.interface.encodeFunctionData('transfer', [
+        deployer.address,
+        1000,
+      ]);
+
       const tx = buildSafeTransaction({
-        to: await mockNFT.getAddress(),
-        data: mintNFTData,
+        to: await votesERC20.getAddress(),
+        data: tokenTransferData,
         safeTxGas: 1000000,
         nonce: await gnosisSafe.nonce(),
       });
@@ -447,9 +389,15 @@ describe('Child Multisig DAO with Azorius Parent', () => {
     });
 
     it('A transaction cannot be executed if the timelock period has not ended yet', async () => {
+      // Create transaction to set the guard address
+      const tokenTransferData = votesERC20.interface.encodeFunctionData('transfer', [
+        deployer.address,
+        1000,
+      ]);
+
       const tx = buildSafeTransaction({
-        to: await mockNFT.getAddress(),
-        data: mintNFTData,
+        to: await votesERC20.getAddress(),
+        data: tokenTransferData,
         safeTxGas: 1000000,
         nonce: await gnosisSafe.nonce(),
       });
@@ -492,16 +440,20 @@ describe('Child Multisig DAO with Azorius Parent', () => {
 
     it('A DAO may execute txs during a the freeze proposal period if the freeze threshold is not met', async () => {
       // Vetoer 1 casts 500 freeze votes
-      await freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
+      await freezeVoting.connect(tokenVetoer1).castFreezeVote();
 
       // Check that the DAO has been frozen
       expect(await freezeVoting.isFrozen()).to.eq(false);
 
+      // Create transaction to set the guard address
+      const tokenTransferData1 = votesERC20.interface.encodeFunctionData('transfer', [
+        deployer.address,
+        1000,
+      ]);
+
       const tx1 = buildSafeTransaction({
-        to: await mockNFT.getAddress(),
-        data: mintNFTData,
+        to: await votesERC20.getAddress(),
+        data: tokenTransferData1,
         safeTxGas: 1000000,
         nonce: await gnosisSafe.nonce(),
       });
@@ -550,9 +502,7 @@ describe('Child Multisig DAO with Azorius Parent', () => {
       expect(await freezeVoting.freezeProposalCreatedBlock()).to.eq(0);
 
       // Vetoer 1 casts 500 freeze votes
-      await freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
+      await freezeVoting.connect(tokenVetoer1).castFreezeVote();
       expect(await freezeVoting.isFrozen()).to.eq(false);
       expect(await freezeVoting.freezeProposalVoteCount()).to.eq(500);
       let latestBlock = await hre.ethers.provider.getBlock('latest');
@@ -561,9 +511,7 @@ describe('Child Multisig DAO with Azorius Parent', () => {
       // Move time forward to elapse freeze proposal period
       await time.advanceBlocks(10);
 
-      await freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
+      await freezeVoting.connect(tokenVetoer1).castFreezeVote();
       expect(await freezeVoting.freezeProposalVoteCount()).to.eq(500);
       latestBlock = await hre.ethers.provider.getBlock('latest');
       expect(await freezeVoting.freezeProposalCreatedBlock()).to.eq(latestBlock!.number);
@@ -571,35 +519,31 @@ describe('Child Multisig DAO with Azorius Parent', () => {
     });
 
     it('A user cannot vote twice to freeze a dao during the same voting period', async () => {
-      await freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
+      await freezeVoting.connect(tokenVetoer1).castFreezeVote();
       await expect(
-        freezeVoting
-          .connect(tokenVetoer1)
-          ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids),
-      ).to.be.revertedWithCustomError(freezeVoting, 'NoVotes()');
+        freezeVoting.connect(tokenVetoer1).castFreezeVote(),
+      ).to.be.revertedWithCustomError(freezeVoting, 'AlreadyVoted');
       expect(await freezeVoting.freezeProposalVoteCount()).to.eq(500);
     });
 
     it('An unfrozen DAO may not execute a previously passed transaction', async () => {
       // Vetoer 1 casts 500 freeze votes
-      await freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
-      // Vetoer 2 casts 1000 freeze votes
-      await freezeVoting
-        .connect(tokenVetoer2)
-        [
-          'castFreezeVote(address[],uint256[])'
-        ]([await mockNFT.getAddress(), await mockNFT.getAddress()], vetoer2Ids);
+      await freezeVoting.connect(tokenVetoer1).castFreezeVote();
+      // Vetoer 2 casts 600 freeze votes
+      await freezeVoting.connect(tokenVetoer2).castFreezeVote();
 
       // Check that the DAO has been frozen
       expect(await freezeVoting.isFrozen()).to.eq(true);
 
+      // Create transaction to set the guard address
+      const tokenTransferData1 = votesERC20.interface.encodeFunctionData('transfer', [
+        deployer.address,
+        1000,
+      ]);
+
       const tx1 = buildSafeTransaction({
-        to: await mockNFT.getAddress(),
-        data: mintNFTData,
+        to: await votesERC20.getAddress(),
+        data: tokenTransferData1,
         safeTxGas: 1000000,
         nonce: await gnosisSafe.nonce(),
       });
@@ -665,24 +609,24 @@ describe('Child Multisig DAO with Azorius Parent', () => {
 
     it('Unfrozen DAOs may execute txs', async () => {
       // Vetoer 1 casts 500 freeze votes
-      await freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
-      // Vetoer 2 casts 1000 freeze votes
-      await freezeVoting
-        .connect(tokenVetoer2)
-        [
-          'castFreezeVote(address[],uint256[])'
-        ]([await mockNFT.getAddress(), await mockNFT.getAddress()], vetoer2Ids);
+      await freezeVoting.connect(tokenVetoer1).castFreezeVote();
+      // Vetoer 2 casts 600 freeze votes
+      await freezeVoting.connect(tokenVetoer2).castFreezeVote();
 
       // Check that the DAO has been frozen
       expect(await freezeVoting.isFrozen()).to.eq(true);
       await freezeVoting.connect(freezeGuardOwner).unfreeze();
       expect(await freezeVoting.isFrozen()).to.eq(false);
 
+      // Create transaction to set the guard address
+      const tokenTransferData1 = votesERC20.interface.encodeFunctionData('transfer', [
+        deployer.address,
+        1000,
+      ]);
+
       const tx1 = buildSafeTransaction({
-        to: await mockNFT.getAddress(),
-        data: mintNFTData,
+        to: await votesERC20.getAddress(),
+        data: tokenTransferData1,
         safeTxGas: 1000000,
         nonce: await gnosisSafe.nonce(),
       });
@@ -729,19 +673,11 @@ describe('Child Multisig DAO with Azorius Parent', () => {
 
     it('You must have voting weight to cast a freeze vote', async () => {
       await expect(
-        freezeVoting
-          .connect(freezeGuardOwner)
-          ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids),
+        freezeVoting.connect(freezeGuardOwner).castFreezeVote(),
       ).to.be.revertedWithCustomError(freezeVoting, 'NoVotes()');
-      freezeVoting
-        .connect(tokenVetoer1)
-        ['castFreezeVote(address[],uint256[])']([await mockNFT.getAddress()], vetoer1Ids);
+      freezeVoting.connect(tokenVetoer1).castFreezeVote();
       await expect(
-        freezeVoting
-          .connect(freezeGuardOwner)
-          [
-            'castFreezeVote(address[],uint256[])'
-          ]([await mockNFT.getAddress(), await mockNFT.getAddress()], vetoer2Ids),
+        freezeVoting.connect(freezeGuardOwner).castFreezeVote(),
       ).to.be.revertedWithCustomError(freezeVoting, 'NoVotes()');
     });
 
@@ -762,7 +698,7 @@ describe('Child Multisig DAO with Azorius Parent', () => {
     });
 
     it('Only the freeze voting owner can update the freeze votes threshold', async () => {
-      expect(await freezeVoting.freezeVotesThreshold()).to.eq(501);
+      expect(await freezeVoting.freezeVotesThreshold()).to.eq(1090);
 
       await freezeVoting.connect(freezeGuardOwner).updateFreezeVotesThreshold(2000);
