+++ test/utilities/DecentHatsModificationModule.test.ts
@@ -30,7 +30,10 @@ import {
   DecentHatsCreationModule,
 } from '../../typechain-types';
 
-import { getGnosisSafeL2Singleton, getGnosisSafeProxyFactory } from '../GlobalSafeDeployments.test';
+import {
+  getGnosisSafeL2Singleton,
+  getGnosisSafeProxyFactory,
+} from '../global/GlobalSafeDeployments.test';
 import {
   executeSafeTransaction,
   getHatAccount,
